+++ ExposurePlayback/Context/ExposureContext+Warning.swift
@@ -35,6 +35,8 @@ extension ExposureContext.Warning {
         case fetchingCurrentProgramFailed(timestamp: Int64, channelId: String, error: ExposureError?)
         case gapInEpg(timestamp: Int64, channelId: String)
         case entitlementValidationFailed(programId: String, channelId: String, error: ExposureError?)
+        case navigateToNextProgramFailed(programId: String, channelId: String, error: ExposureError?)
+        case navigateToPreviousProgramFailed(programId: String, channelId: String, error: ExposureError?)
     }
 }
 
@@ -44,6 +46,8 @@ extension ExposureContext.Warning.ProgramService {
         case .fetchingCurrentProgramFailed(timestamp: let timestamp, channelId: let channelId, error: let error): return "Program Service failed to fetch the current program at timestamp \(timestamp) on \(channelId). errorCode: \(error?.code) message: \(error?.message)"
         case .gapInEpg(timestamp: let timestamp, channelId: let channelId): return "Program Service encountered a gap in the Epg at timestamp \(timestamp) on \(channelId)"
         case .entitlementValidationFailed(programId: let programId, channelId: let channelId, error: let error): return "Program Service failed to validate program \(programId) on \(channelId). errorCode: \(error?.code) message: \(error?.message)"
+        case .navigateToNextProgramFailed(programId: let programId, channelId: let channelId, error: let error): return "Navigating to next program is not allowed \(programId) on \(channelId). errorCode: \(error?.code) message: \(error?.message)"
+        case .navigateToPreviousProgramFailed(programId: let programId, channelId: let channelId, error: let error): return "Navigating to previous program is not allowed \(programId) on \(channelId). errorCode: \(error?.code) message: \(error?.message)"
         }
     }
 }