+++ ExposurePlayback/Services/ProgramService.swift
@@ -12,6 +12,8 @@ import Player
 
 internal protocol ProgramProvider {
     func fetchPrograms(on channelId: String, timestamp: Int64, using environment: Environment, callback: @escaping ([Program]?, ExposureError?) -> Void)
+    func fetchNextProgram(on program: Program, using environment: Environment, callback: @escaping (Program?, ExposureError?) -> Void)
+    func fetchPreviousProgram(on program: Program, using environment: Environment, callback: @escaping (Program?, ExposureError?) -> Void)
     func validate(entitlementFor assetId: String, environment: Environment, sessionToken: SessionToken, callback: @escaping (EntitlementValidation?, ExposureError?) -> Void)
 }
 
@@ -111,6 +113,41 @@ public class ProgramService {
     
     internal var provider: ProgramProvider
     internal struct ExposureProgramProvider: ProgramProvider {
+        
+        
+        /// Fetch next program
+        ///
+        /// - Parameters:
+        ///   - program: current program
+        ///   - environment: exposure environment 
+        ///   - callback: callback to fire once the event is fired.
+        func fetchNextProgram(on program: Program, using environment: Environment, callback: @escaping (Program?, ExposureError?) -> Void) {
+            FetchEpg(environment: environment)
+                .next(programId: program.programId)
+                .request()
+                .validate()
+                .response {
+                    callback($0.value, $0.error)
+            }
+        }
+        
+        
+        /// Fetch the previous program
+        ///
+        /// - Parameters:
+        ///   - program: current program
+        ///   - environment: exposure enviornment
+        ///   - callback: callback to fire once the event is fired.
+        func fetchPreviousProgram(on program: Program, using environment: Environment, callback: @escaping (Program?, ExposureError?) -> Void) {
+            FetchEpg(environment: environment)
+                .previous(programId: program.programId)
+                .request()
+                .validate()
+                .response {
+                    callback($0.value, $0.error)
+            }
+        }
+        
         func fetchPrograms(on channelId: String, timestamp: Int64, using environment: Environment, callback: @escaping ([Program]?, ExposureError?) -> Void) {
             FetchEpg(environment: environment)
                 .channel(id: channelId)
@@ -118,7 +155,9 @@ public class ProgramService {
                 .filter(onlyPublished: true)
                 .request()
                 .validate()
-                .response{ callback($0.value?.programs, $0.error) }
+                .response{
+                    callback($0.value?.programs, $0.error)
+            }
         }
         
         
@@ -148,25 +187,71 @@ extension ProgramService {
     }
     
     fileprivate func fetchProgram(timestamp: Int64, callback: @escaping (Program?, ExposureContext.Warning.ProgramService?) -> Void) {
-        provider.fetchPrograms(on: channelId, timestamp: timestamp, using: environment) { [weak self] newPrograms, error in
-            guard let `self` = self else { return }
-            guard error == nil else {
-                // There was an error fetching the program. Be permissive and allow playback
-                callback(nil, .fetchingCurrentProgramFailed(timestamp: timestamp, channelId: self.channelId, error: error))
-                return
-            }
+        
+        self.currentProgram(for: timestamp, callback: { program , error in
             
-            if let programs = newPrograms, let program = self.requestedProgram(for: timestamp, fromCandidates: programs) {
-                callback(program, nil)
-            }
-            else {
-                /// Validation on program level requires the channel has Epg attached.
-                ///
-                /// If we are missing Epg, playback is allowed to continue.
+            if let currentlyPlayingProgram = program {
+                
+                // If the currenet program end time & the fetched program endtime is different :=>
+                // Current program has extended
+                if currentlyPlayingProgram.endDate?.millisecondsSince1970 != self.currentProgram?.endDate?.millisecondsSince1970 {
+                    callback(program, nil)
+                } else {
+                    //// Using new NextProgram endpoint to get the next Protram
+                    self.provider.fetchNextProgram(on: currentlyPlayingProgram, using: self.environment, callback: { [weak self ] nextProgram, error in
+                        guard let `self` = self else { return }
+                        guard error == nil else {
+                            // There was an error fetching the program. Be permissive and allow playback
+                            callback(nil, .fetchingCurrentProgramFailed(timestamp: timestamp, channelId: self.channelId, error: error))
+                            return
+                        }
+                        if nextProgram != nil {
+                            if nextProgram?.startDate?.millisecondsSince1970 != currentlyPlayingProgram.endDate?.millisecondsSince1970 {
+                                // EPG HAS A GAP (nextprogram start time is not equal to the current program's end time )
+                                callback(nil, .gapInEpg(timestamp: timestamp, channelId: self.channelId))
+                            } else {
+                                callback(nextProgram, nil)
+                            }
+                        }
+                        else {
+                            // GAP in EPG
+                            callback(nil, .gapInEpg(timestamp: timestamp, channelId: self.channelId))
+                        }
+                    })
+                    
+                    // PRevious implementation using timestamps
+                    /* self.provider.fetchPrograms(on: self.channelId, timestamp: timestamp, using: self.environment) { [weak self] newPrograms, error in
+                        guard let `self` = self else { return }
+                        guard error == nil else {
+                            // There was an error fetching the program. Be permissive and allow playback
+                            callback(nil, .fetchingCurrentProgramFailed(timestamp: timestamp, channelId: self.channelId, error: error))
+                            return
+                        }
+                        
+                        if let programs = newPrograms, let program = self.requestedProgram(for: timestamp, fromCandidates: programs) {
+                            callback(program, nil)
+                        }
+                        else {
+                            /// Validation on program level requires the channel has Epg attached.
+                            ///
+                            /// If we are missing Epg, playback is allowed to continue.
+                            callback(nil, .gapInEpg(timestamp: timestamp, channelId: self.channelId))
+                        }
+                    } */
+                    
+                }
+            } else {
+                // When running the unit tests current program can be nil
+                guard error == nil else {
+                    // There was an error fetching the program. Be permissive and allow playback
+                    callback(nil, .fetchingCurrentProgramFailed(timestamp: timestamp, channelId: self.channelId, error: error))
+                    return
+                }
                 callback(nil, .gapInEpg(timestamp: timestamp, channelId: self.channelId))
             }
-        }
+        })
     }
+    
     
     fileprivate func validate(program: Program, callback: @escaping (ExposureContext.Warning.ProgramService?, String?) -> Void) {
         self.provider.validate(entitlementFor: program.assetId, environment: self.environment, sessionToken: self.sessionToken) { [weak self] validation, error in
@@ -188,13 +273,22 @@ extension ProgramService {
         }
     }
     
-    internal func handleProgramChanged(program: Program?) {
+    internal func handleProgramChanged(program: Program?, isExtendedProgram: Bool) {
         DispatchQueue.main.async { [weak self] in
             guard let `self` = self else { return }
             let current = self.activeProgram
             self.activeProgram = program
-            if current?.programId != program?.programId {
+            if current?.programId != program?.programId && isExtendedProgram == false {
                 self.onProgramChanged(program)
+            }
+            
+            // The current program has extended, call program changed even if it's the same program
+            else if ( current?.programId == program?.programId && isExtendedProgram == true ) {
+                self.onProgramChanged(program)
+            }
+            
+            else {
+                // Do nothing
             }
         }
     }
@@ -266,23 +360,23 @@ extension ProgramService {
         stopProgramChangeTimer()
         
         guard let timestamp =  currentPlayheadTime() else { return }
+        
         provider.fetchPrograms(on: channelId, timestamp: timestamp + monitoringOffset(), using: environment) { [weak self] newPrograms, error in
             guard let `self` = self else { return }
             guard error == nil else {
                 // We are permissive on errors, allow playback
-                self.handleProgramChanged(program: nil)
+                self.handleProgramChanged(program: nil, isExtendedProgram: false)
                 self.onWarning(.fetchingCurrentProgramFailed(timestamp: timestamp, channelId: self.channelId, error: error))
                 return
             }
             
             guard let programs = newPrograms, let program = self.requestedProgram(for: timestamp, fromCandidates: programs) else {
                 // There is no program, validation can not occur, allow playback
-                self.handleProgramChanged(program: nil)
+                self.handleProgramChanged(program: nil, isExtendedProgram: false)
                 self.onWarning(.gapInEpg(timestamp: timestamp, channelId: self.channelId))
                 return
             }
-            
-            self.handleProgramChanged(program: program)
+            self.handleProgramChanged(program: program, isExtendedProgram: false)
             self.startValidationTimer(onTimestamp: timestamp, for: program)
         }
     }
@@ -320,7 +414,10 @@ extension ProgramService {
                                 self.onWarning(warning)
                             }
                             // Trigger the program change no matter what
-                            self.handleProgramChanged(program: program)
+                            // If the program has extended, it needs to be handled : isExtendedProgram => true
+                            let isextended = program?.programId == self.currentProgram?.programId ? true : false
+                            
+                            self.handleProgramChanged(program: program, isExtendedProgram: isextended)
                             
                             if let program = program {
                                 // If we received a program, validate that
@@ -382,5 +479,42 @@ extension ProgramService {
             
             callback(program, nil)
         }
+    }
+}
+
+// MARK: - Next & Previous Program
+extension ProgramService {
+    
+    /// Call the provider for next program
+    ///
+    /// - Parameters:
+    ///   - program: current program
+    ///   - callback: callback: callback to fire once the event is fired.
+    func nextProgram(program: Program, callback: @escaping (Program?, ExposureError?) -> Void) {
+        
+        let currentTimeStamp = Date().millisecondsSince1970
+        guard let start = currentProgram?.startDate?.millisecondsSince1970, let end = currentProgram?.endDate?.millisecondsSince1970 else { return }
+        if start <= currentTimeStamp && currentTimeStamp < end {
+            let error = NSError(domain: "This is a live Program, navigating to the next program is not allowed", code: 40, userInfo: nil)
+            self.onWarning(.navigateToNextProgramFailed(programId: program.programId, channelId: channelId, error: ExposureError.generalError(error: error)))
+            callback(nil, nil)
+        }
+        else {
+            provider.fetchNextProgram(on: program, using: environment, callback: { nextProgram, error in
+                callback(nextProgram, error)
+            })
+        }
+    }
+    
+    
+    /// Call the provider for previous program
+    ///
+    /// - Parameters:
+    ///   - program: current program
+    ///   - callback: callback: callback to fire once the event is fired.
+    func previousProgram(program: Program, callback: @escaping (Program?, ExposureError?) -> Void) {
+        provider.fetchPreviousProgram(on: program, using: environment, callback: { previousProgram, error in
+            callback(previousProgram, error)
+        })
     }
 }